services/tiktok-transcriber/requirements.txt
Flask==3.1.2
Flask-Cors==6.0.1
gunicorn==23.0.0
requests==2.32.5
google-generativeai==0.8.5
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
gspread==6.2.1
pandas==2.3.2
openpyxl==3.1.5

